feat(673): refactor utils functions out of index file
