+++ packages/frontend-asset-loader/src/__tests__/index.test.ts
@@ -1,4 +1,4 @@
-import { parseAssets } from "../index";
+import { parseAssets } from "../utils/utils";
 
 describe("parseAssets", () => {
   const input = [
