+++ packages/frontend-asset-loader/src/index.ts
@@ -3,9 +3,10 @@ import fg from "fast-glob";
 import { Express } from "express";
 import {
   getDuplicateHashedFileName,
-  isValidHashName,
+  mapAssetsToLocal,
+  parseAssets,
 } from "./utils/utils";
-import { PathAndFile } from "./utils/utils.types";
+
 import { getLogger, CustomLogger, setLogger } from "./utils/logger";
 
 export const loadAssets = (
@@ -29,25 +30,3 @@ export const loadAssets = (
   mapAssetsToLocal(app, pathsAndFiles);
 };
 
-export const parseAssets = (assets: string[], hashBetween: { start: string; end: string }): PathAndFile[] => {
-  return assets.map((asset) => {
-    const pathParts = asset.split("/");
-    const hashedFileName = pathParts[pathParts.length - 1];
-
-    if (isValidHashName(hashedFileName, hashBetween.start, hashBetween.end)){
-      
-      const [fileName, hashedExtension] = hashedFileName.split(hashBetween.start);
-      const [hash, extension] = hashedExtension.split(hashBetween.end);
-      return { hashedFileName, hash: hash, fileName: `${fileName}.${extension}`};
-    }
-
-    return { hashedFileName, fileName: hashedFileName };
-  });
-};
-
-export const mapAssetsToLocal = (app: Express, pathsAndFiles: PathAndFile[]) => {
-  pathsAndFiles.forEach((pathAndFile) => {
-    app.locals.assets = app.locals.assets || {};
-    app.locals.assets[pathAndFile.fileName] = pathAndFile.hashedFileName;
-  });
-};
