+++ packages/frontend-asset-loader/src/utils/utils.ts
@@ -1,5 +1,12 @@
 import { PathAndFile } from "./utils.types";
 import { getLogger } from "../utils/logger";
+import { Express } from "express";
+
+declare module 'express' {
+  interface Application {
+    locals: { [key: string]: unknown };
+  }
+}
 
 export const getDuplicateHashedFileName = (
   array: PathAndFile[],
@@ -52,3 +59,30 @@ export const isValidHashName = (
 
   return true;
 };
+
+
+export const parseAssets = (assets: string[], hashBetween: { start: string; end: string }): PathAndFile[] => {
+  return assets.map((asset) => {
+    const pathParts = asset.split("/");
+    const hashedFileName = pathParts[pathParts.length - 1];
+
+    if (isValidHashName(hashedFileName, hashBetween.start, hashBetween.end)){
+      
+      const [fileName, hashedExtension] = hashedFileName.split(hashBetween.start);
+      const [hash, extension] = hashedExtension.split(hashBetween.end);
+      return { hashedFileName, hash: hash, fileName: `${fileName}.${extension}`};
+    }
+
+    return { hashedFileName, fileName: hashedFileName };
+  });
+};
+
+export const mapAssetsToLocal = (app: Express, pathsAndFiles: PathAndFile[]) => {
+  app.locals = app.locals || {}; 
+
+  pathsAndFiles.forEach((pathAndFile) => {
+    app.locals.assets = app.locals.assets || {};
+    app.locals.assets[pathAndFile.fileName] = pathAndFile.hashedFileName;
+  });
+};
+
